chore: Bump font-awesome version
The way of integrating font-awesome into the project is via a script tag
now instead of the usual link tag. They are using bundles, called "kits"
tailored for each user. When a user creates a kit, a script is created for
them which is fetched via their domain and loaded into your app.

Resolves #199